Load pending resources after style adjustments.
We don't fetch resources for ComputedStyle which is in a display:none
subtree or display:contents since [1]. Style adjustments are done to the
display type in the StyleAdjuster and we need to make sure we load
resources where display changes from a non-rendered value to a rendered
one after style adjustments.

Specifically, this happens with blockification of :root elements which
are display:contents.

[1] https://crrev.com/caf740f6c0b57ca102026f40f3cc62f9f3cf5159

Bug: 1085750
Change-Id: I7686c6f8bf4918bec5812d1bf893ad2c4cb9262c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2214897
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#771748}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 19e420c62611f97548088bf9f3b3d480f4ba4014