[EventTiming] Fix duration value
The WPT.FYI results [1] show that several tests which pass on our bots
are actually timing out there. I dug into this and it seems to be due to
improper usage of swap promises. In the testcase I looked, a mouseover
event occurs, triggering the swap promise. Then a long mousedown occurs.
While it is being processed, the GPU completes the frame and resolves
the swap promise. However, the main thread is busy at the moment while
its handling the mousedown, so it does not yet know about this. In the
current logic, there is still a pending swap promise, hence no new swap
promise is queued when mousedown handlers have finished executing. This
means that both the mouseover and mousedown events use the same swap
promise, one which had a paint timestamp which is way before the
mousedown handlers have even finished. This means that the |duration|
for the mousedown entry is also much less, and hence no entry is
dispatched for this event and the tests timeout.

The fix is to queue a swap promise when none are queued, or when a new
paint lifecycle occurs. This way we queue at most one promise per frame
when we're actively painting content, and have at most one queued
promise when the renderer is not painting. This fixes the tests, at
least per my manual running of /.wpt run, which reproduces the timeouts
seen in wpt.fyi.

[1 ]https://wpt.fyi/results/event-timing?label=experimental&label=master&aligned

Bug: 543598
Change-Id: I298abb1e13adb0b17f1b2d493cb18b43c407c863
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2174880
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Reviewed-by: Steve Kobes <skobes@chromium.org>
Commit-Queue: Nicolás Peña Moreno <npm@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#771773}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fe961ee51f0f089a06dba66d4a5690dfdf2da5b5