Move (and rewrite) pointer lock implementation to EventSender
Pointer locking is a property of the widget, not a specific frame, nor
the process. So we move it to the EventSender which is the per-widget
web test javascript interface.

This removes the last usage of TestRunnerForSpecificView, and we remove
the class and all pointers to it.

The pointer lock implementation is made a bit more clear and robust,
tracking the state of pending locks/unlocks, as well as the final
locked state. We use the default task runner, as these actions are
tied to the widget and IPC, rather than to the requesting frame.

R=avi@chromium.org
TBR=dcheng

Bug: 866140, 1069111
Change-Id: I90e6bfd336b4647fdead185012bf9b84b6617af2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2212928
Commit-Queue: danakj <danakj@chromium.org>
Auto-Submit: danakj <danakj@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#771869}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3dec984312e13fa5f9543df7beb481f42b2cc258