ZIndex: Report computed z-index as specified, ignoring stacking context.
This patch changes the behavior of computed style z-index to
comply with spec: it reports the value as specified instead of
changing it to auto if the style does not induce a stacking context.

Note that since z-index only applies on positioned elements, and
grid and flex children, this patch introduces an EffectiveZIndex() which
is set to 0 if the z-index does not apply. This is so that code relying
on z-index will behave as it did before. However, ZIndex() now reflects
the value that was specified.

R=futhark@chromium.org, pdr@chromium.org

Fixed: 1023232, 375982
Change-Id: Ib11467a9818b746c72bebbb167223f1bb41a5f9f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2212391
Auto-Submit: vmpstr <vmpstr@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: vmpstr <vmpstr@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#771989}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 005471b2123dc5fcfb223042b89ddc486e7df7ea