Use light instead of no-preference for printing.
We use a preferred color-scheme of light for screens. When we decided to
not use dark for printing we chose no-preference, but content depends on
color-scheme to either be light or dark since that is the case for
screen media. This matches what both Firefox and Safari does.

Bug: 1086441
Change-Id: I1b10c76068ca9cfebf0d991c342b15f99185698e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2216770
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#771997}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 44b77b6b52bd86030f1f49475fc04d9d1a3311d4