Perform monospace-related adjustment using true parent style
There is currently some code in FontBuilder that tweaks the specified
font size if the current generic font family changes to or from
monospace. This code assumes that the FontDescription stored on the
ComputedStyle pre-CreateFont is the inherited FontDescription, but
this is no longer (always) true.

This means that declarations like font-size:1em can be applied (and
subjected to CreateFont) twice, which means that the second call to
CheckForGenericFamilyChange won't detect a change in the
monospace-ness, hence the specified size will fail to adjust.

This CL fixes this by propagating the actual parent style to
CheckForGenericFamilyChange.

Bug: 1086082, 1086680
Change-Id: Ia61a2327890f27497e4ea5f41dec00a5e0a450d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2216511
Reviewed-by: Xiaocheng Hu <xiaochengh@chromium.org>
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#772032}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ca7df536bf7bc572825fe4efba2766b34c092698