Fixed issue when the focus ring is clipped for textarea.
Prior to http://crrev.com/c/2138156, the outline-offset was ignored
when drawing the focus ring.

Unlike normal outlines (whole width is outside of the offset), focus
rings are drawn with the center of the path aligned with the offset, so
only a part of the ouline is outside of the offset.

This CL updates the behavior of NeedsDecorationOutlineLayer to consider
the part of the outline drawn inside.

Added textarea-with-scrollbar.html to validate the change.

Bug: 1084997
Change-Id: I50949194c0a2f413d7a3c508509339dac63cfbb6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2216705
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Commit-Queue: Ionel Popescu <iopopesc@microsoft.com>
Cr-Original-Commit-Position: refs/heads/master@{#772103}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ceeea9346811da72827183a3bd126f844bd12c2c