Add IChromeAccessible, a custom async COM accessibility API
This change introduces the first draft of a possible
Chrome-specific COM API for accessibility that's async-only
and could potentially lead to better performance and
more easily allow Windows assistive technology to avoid
injecting in-process.

This new interface is behind a feature flag with no intention
of launching as-is - if this proves to be successful we
should either standardize this as part of an existing
API like IAccessible2, or else create a new public spec
for it.

For now, the goal is to make something available for
experimenting and building a proof-of-concept.

Bug: 1083834
AX-Relnotes: None
Change-Id: I32fb830217f55e1b1b214b1044362d9c2c42f5b7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2206224
Reviewed-by: Nico Weber <thakis@chromium.org>
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Commit-Queue: Dominic Mazzoni <dmazzoni@chromium.org>
Auto-Submit: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#772425}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b1cdadd9c32c7222bd4a498d46a52dc446589e1c