New Windows root scroller overscroll - Part 1.
This CL allows the root scroller to be overscrolled beyond the scroll
bounds. The OverscrollBounceController drives this. The basic principle
here is: the scroll deltas on GSU are fed into a hyperbolic tangent
function (ranging [0, 2]). This tells how far the scroller bounds can
be "stretched" when the user overscrolls. See OverscrollBounceDistance
for details. Animating the bounce back is not in the scope of this CL.

Bug: 1058071
Change-Id: I3023c69078ffc6dff0ecfe3554d699ea3b9d1553
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2198596
Commit-Queue: Rahul Arakeri <arakeri@microsoft.com>
Reviewed-by: David Bokan <bokan@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#772484}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a675fe0f941b03ae443027663aa41f8eef13f401