Fix the http/tests/devtools/inspect-iframe-from-different-domain.js test
This test wants to verify devtools inspect works with an iframe in
another domain. It was putting the iframe in another site, which causes
it to become detached from the local root in site isolation. Instead,
put both the main frame and iframe into the example.test domain, with
the iframe in a subdomain.

This test was originally added for webkit bug 31587 which is found at
https://bugs.webkit.org/show_bug.cgi?id=31587.

It said:
> If you try to inspect elements located inside an IFRAME that was
> loaded from a different domain, not satisfying the same origin
> policy, Web Inpector in recent WebKit nightly builds cannot show
> style information. Works in Safari 4.0.4.

R=yangguo@chromium.org

Bug: 623268
Change-Id: I842bf835e49ee79e60e423fd36a5d3cedb8a8cfa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2218765
Commit-Queue: danakj <danakj@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#772743}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 002510a0ea36ea60e771efbf897428f9487c5ed2