Fix the http/tests/devtools/console-cross-origin-iframe-logging.js test
This test wants to verify that an error is seen when using a frame in
another origin. It was using a different site, where the frame's
contents is not visible at all when site-isolated. Instead, put the
frame in another origin in the same site so that the main frame can see
the contents of the iframe.

R=caseq@chromium.org

Bug: 623268
Change-Id: Ib4ab03229f1033fc01717e561eb35b3c6d047e2c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2218373
Commit-Queue: danakj <danakj@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#772769}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d18bad6451449792d5bba79b0331b86fb9581e6e