[ScrollUnification] Teach CC to request Blink hit test
This CL builds on https://crrev.com/c/2199610 where we added a new
mechanism by which the compositor can respond to an event by requesting
a main thread hit test. See that CL description for details on how the
eventing side of this works. This CL makes use of this new machinery and
implements the compositor hit testing and scrolling required to perform
scrolls of uncomposited/un-hit-testable/mainThreadReasoned scrollers
from the compositor thread.

The compositor will now respond with the |needs_main_thread_hit_test|
bit set when ScrollBegin finds a situation it cannot hit test on its
own (i.e. where it previously would have sent the scroll to the main
thread). In this case, the client will post a task to Blink to perform a
hit test. When Blink finishes, the client will call ScrollBegin again,
with a hit-tested ElementId and a bit stating that the hit test has been
provided from the main thread.

We also make it so that only the ScrollNode is mutated when scrolling an
uncomposited (no cc::Layer) or main-thread-scrolling-reason scroller
(composited but requires repaint-on-scroll or some other main thread
synchronization). Scrolling these kinds of scrollers from the compositor
will ensure the TransformNode is untouched. For an uncomposited
scroller, mutating the transform node is a harmless no-op as there is no
layer so this won't affect what's outputted by a draw.  However, for
MTSR scrollers doing so can lead to incorrect scrolling artifacts.  Such
scrollers do have a cc::Layer but it expects never to be scrolled from
the compositor. In these cases, we want to update the ScrollNode so that
we send an updated scroll offset to Blink in the next BeginMainFrame.
When that BMF commits, we'll receive updated property trees and
repainted layers. This is how "main thread" scrolls are actualized in
scroll unification.

Finally, we disable the main thread scroll handling paths when
unification is turned on to make it easy to see that scrolls are indeed
occurring from the compositor thread and we aren't accidentally falling
back to the main thread.

Bug: 1047182
Change-Id: I7eba93488359f43fb15f0604f1fd3032bd4f910a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2189580
Reviewed-by: Robert Flack <flackr@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: David Bokan <bokan@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#772786}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1061035574c5def3b8288db856cfe98329b7f735