Avoid repeating a single row/column of tiled background images
When working in subpixel coordinates background image tiles can end up
slightly off integral coordinates when they should be exactly integral.
When the image needs to repeat in one dimension but not the other, this
slight mismatch can lead to a single row/column of pixels in the
non-repeating direction being incorrectly repeated.

To avoid this, detect the situation and force the coordinates to
integral values. This is done as far up the stack as possible so
that we can be sure to enlarge the tile enough to fill any gaps
to borders.

Bug: 1085801
Change-Id: I9de60ce3c6f6e057550dd590c15c1554ea9f36d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2219353
Commit-Queue: Stephen Chenney <schenney@chromium.org>
Reviewed-by: Fredrik Söderquist <fs@opera.com>
Cr-Original-Commit-Position: refs/heads/master@{#772880}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6956d3dac8f9ff1208ce915b75641722b26a5454