Remove VTT usages of CreateElementFlags::ByParser
We're interesting in implementing the HTML Standard notion of a parser
document [1], which will replace CreateElementFlags::created_by_parser_
with a pointer to the document that the parser was created for.
Therefore when marking a CreateElementFlags object as created-by-parser,
one should also provide the document for the parser.

As a part of this we're looking at all of the call-sites of
CreateElementByFlags::{ByParser(),SetCreatedByParser()} to see which are
actually related to parsing an HTML Document and require parser-inserted
information.

It seems that the VTT usages of CreateElementFlags::ByParser is
unnecessary as it doesn't have anything to do with parsing an HTML
document, and the parser-inserted information is never not queried for
elements created here.

[1]: https://html.spec.whatwg.org/C/#parser-document

R=hiroshige@chromium.org, kouhei@chromium.org

Bug: 1086507
Change-Id: Id40ddc2f7d87652cee8d6f2832e307bdf8f05d46
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2216314
Reviewed-by: Mason Freed <masonfreed@chromium.org>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Reviewed-by: Hiroshige Hayashizaki <hiroshige@chromium.org>
Commit-Queue: Dominic Farolino <dom@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#772972}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 93c36bef6744b6874c558dd7c0c490d5210547cb