DCHECK and TODO for crbug.com/786475 (Fieldset should not scroll)
This makes the failure more explicit, and make the expectation aligned
in pre-CompositeAfterPaint and CompositeAfterPaint for
external/wpt/html/rendering/non-replaced-elements/the-fieldset-and-legend-elements/fieldset-overflow.html.

Previously in CompositeAfterPaint the test triggered a DCHECK that was
not obvious of the reason.

Bug: 786475
Change-Id: I93c7df7c731b9f87d88d5ad509d2c6a4d4b10755
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2220159
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#773249}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1239a45d2cd8823f6ea1db62d29fa24099763668