Fix persistent active user gesture on RemoteFrames after one activation
If a local frame gets a user activation, we set a bit on FrameTreeNode,
indicating that it has once received a user gesture. That bit is then
used to set an /active/ user gesture in RemoteFrames created for the
FrameTreeNode thereafter.

This manifests as flaky web tests that see a user gesture in OOP
iframes that they were not expecting - because they see an active user
gesture in the remote main frame that was created and consumed in the
previous test.

This would manifest in the wild as an OOPIF gaining a user gesture
that it can use to autoplay or whatnot when there is no active user
gesture.

R=avi@chromium.org, beccahughes@chromium.org, dcheng@chromium.org

Bug: 872952
Change-Id: Ia6f76ab33f4906e5ed60c553a5e9ea07ad75b439
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2220475
Commit-Queue: Daniel Cheng <dcheng@chromium.org>
Auto-Submit: danakj <danakj@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Becca Hughes <beccahughes@chromium.org>
Reviewed-by: Mustaq Ahmed <mustaq@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#773323}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 359a434607adf5ca1c3fbe228a3f79e071fd1e30