Fix http/tests/loading/bad-server-subframe.html for site isolation.
Provisional frames don't have a name, so the logging is different, but
we can print the name of the frame they will be replacing. That name is
assigned at swap in, but the tests can get to it before-hand.

Avoid sending a DidFailProvisionalLoad() notification during unload
of a frame being swapped out. We were sending this unconditionally on
unload even if the navigation succeeded, because the frame being
unloaded has no idea. The navigation is happening in the provisional
frame. This makes the RenderFrameObserver expectations match with site
isolation and without.

R=dgozman@chromium.org

Bug: 765779
Change-Id: I2c160091e06a890470746356b6fcfed3cc977948
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2219088
Auto-Submit: danakj <danakj@chromium.org>
Reviewed-by: Dmitry Gozman <dgozman@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: danakj <danakj@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#773384}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 654897c9cf59c90c316fffb9fdce043b6df2c08f