Process changed children and text with clean layout
This adds some DCHECKs and cleans up some TODOs that were left from
the work to ensure accessibility processing occurs with a clean layout.

Fixing these items required additional changes to active descendant,
so that the serialized active descendant id for visible target nodes.
Without that change, an active descendant made visible did not result
in a change to the id, and therefore no event was fired.
This was due to a change in the order and timing of related code paths.
Another active descendant change was required on Linux to deal with
focusing a container with the active descendant after the fact.

This cleanup is a necessary precursor to the work in CL:2204154.

AX-Relnotes: n/a
Bug: None
Change-Id: I2194bb65bfb8aef9c1c07152dc8b6f144384c74d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2204422
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Reviewed-by: Alice Boxhall <aboxhall@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#773510}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a3abc606149c0ea558edde48c0c1586ef51b05f8