StackingContext: Optimize IsStackingContext/IsStacked checks.
This patch removes duplication in the containment checks when checking
for IsStackingContext. As well, it flips the IsStacked checks since
it is cheaper to check for positioned element and is also a more
common reason for being stacked.

R=pdr@chromium.org

Bug: 1087195, 1087259, 1087526
Change-Id: Ife728dd48512d04ee489a2acf96291c93c30d52a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2222880
Commit-Queue: vmpstr <vmpstr@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#773677}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: aa574e64f9a84604a5af609a2343f04a5649c165