[bfcache] Blocklist SpeechRecognizer
This CL disables BackForwardCache when SpeechRecognizer API is used,
as we believe that currently we do not handle the speech recognition
properly when the page is put into BackForwardCache.

Supporting the SpeechRecognizer feature is to be compatible with
BackForwardCache is the next step.

BUG=1088445

Change-Id: I85334ac00251a1ead69f26c5de0e2f76e07066db
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2224981
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Commit-Queue: Sreeja Kamishetty <sreejakshetty@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#773726}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: dcb4eca88e8c305b59eff9709ea9d40348ac19cb