Fix navigation-blocking-xorigin-iframe.js to not be flaky.
The test wants to wait for the frames to all load, but with site
isolation we have an extra load start/stop on the frame being swapped
out, so it can't use a hard-coded number of loads.

Instead, we wait for the known number of network requests to be made,
and also wait for all the frames to finish loading by watching the
number which have started loading. That brings us to a steady state at
which we can grab a non-flaky snapshot.

R=caseq@chromium.org

Bug: 906879
Change-Id: Id6458313fc4c946f2327af26b8ea81478861ec5f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2222726
Commit-Queue: Andrey Kosyakov <caseq@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Auto-Submit: danakj <danakj@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#773732}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 378036cdca8c12c4c1ec2b2f2ac02e1be2d3a4e3