Remove WebNavigationControl::IsClientNavigationInitialHistoryLoad
With PerNavigationMojoInterface launched, this boolean is not needed.

Initial history navigation in the child frame will be canceled by
destroying the NavigationClient. CommitNavigation IPC will be dropped
on the floor because there is no receiver in the renderer, and browser
will react to the NavigationClient disconnect and abandon the
navigation.

Bug: 1020175
Change-Id: Ifddd2d74ca2897283c39e47ec0f57d3a7470f72c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2215380
Reviewed-by: Arthur Hemery <ahemery@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Dmitry Gozman <dgozman@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#773760}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 090e3639cd76073c4633da46ec9c4f8d7e1182d2