RubyNG: Make ScrollableOverflow compatible with the legacy layout
Before this CL ScrollableOverflow in LayoutNG was always larger than the
line's bounding box.  It was not compatible with the legacy layout, and
it's helpless for <rt> positioning.  This CL makes ScrollableOveflow
compatible with the legacy layout.

Bug: 1069817
Change-Id: I876874966088a59fdc218a83342c1d0e4c68ee87
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2225685
Commit-Queue: Koji Ishii <kojii@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#774069}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5cebb76f6f011767fa81ddd4ec0995251dd5bd18