[throughput] Report median throughput to UKM
Right now we report throughput to UKM system at:
  1. 5 seconds after a sequence starts.
  2. Every 500 seconds after that as long as the sequence lasts.
The problem is that in many cases users stay on a page less than 500
seconds, and thus we would only get the report for the first 5 seconds,
which can be heavily dominated by the page loading period.

This CL changes it such that we record the throughput every 5 seconds,
and then report the median throughput to the UKM when the page shuts
down. At this moment, this applies to the kUniversal tracker only,
which tracks all kinds of frame productions. For other types of
trackers, such as the trackers tracks animation/scroll, etc, this
change will be applied later.

Bug: 1040634
Change-Id: I452e16c45187393f29abf1bdf3650a4e5ffb45fc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2112953
Reviewed-by: Dmitry Gozman <dgozman@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Bryan McQuade <bmcquade@chromium.org>
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Reviewed-by: danakj <danakj@chromium.org>
Commit-Queue: Xida Chen <xidachen@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#774242}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a6f72d2d1cd60d5e2b1dc80d75187a9ab2abac7a